feat: added command line option for dpkg-status

+++ cmd/download.go
@@ -27,6 +27,11 @@ func download(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
+	dpkgStatus, err := cmd.Flags().GetString("dpkg-status")
+	if err != nil {
+		return err
+	}
+
 	downloadDir, err := cmd.Flags().GetString("download-dir")
 	if err != nil {
 		return err
@@ -37,36 +42,38 @@ func download(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	f, err := os.Open("testdata/var/lib/dpkg/status")
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-
-	packageStatuses, err := dpkg.ParsePackageStatuses(f)
-	if err != nil {
-		return err
-	}
-
-	dirEntries, err := os.ReadDir(downloadDir)
-	if err != nil {
-		return err
-	}
-
-	for _, dirEntry := range dirEntries {
-		if dirEntry.IsDir() {
-			continue
+	if dpkgStatus != "" {
+		f, err := os.Open(dpkgStatus)
+		if err != nil {
+			return err
 		}
-		if !strings.HasSuffix(dirEntry.Name(), ".deb") {
-			continue
+		defer f.Close()
+
+		packageStatuses, err := dpkg.ParsePackageStatuses(f)
+		if err != nil {
+			return err
 		}
-		parts := strings.Split(dirEntry.Name()[0:len(dirEntry.Name())-4], "_")
-		packageStatus := packageStatuses.FindPackageStatus(parts[0])
-		if (packageStatus != nil) && (packageStatus.Status == "install ok installed") {
-			fmt.Printf("%s exists\n", packageStatus.Package)
-			err = os.Remove(filepath.Join(downloadDir, dirEntry.Name()))
-			if err != nil {
-				return err
+
+		dirEntries, err := os.ReadDir(downloadDir)
+		if err != nil {
+			return err
+		}
+
+		for _, dirEntry := range dirEntries {
+			if dirEntry.IsDir() {
+				continue
+			}
+			if !strings.HasSuffix(dirEntry.Name(), ".deb") {
+				continue
+			}
+			parts := strings.Split(dirEntry.Name()[0:len(dirEntry.Name())-4], "_")
+			packageStatus := packageStatuses.FindPackageStatus(parts[0])
+			if (packageStatus != nil) && (packageStatus.Status == "install ok installed") {
+				fmt.Printf("%s exists\n", packageStatus.Package)
+				err = os.Remove(filepath.Join(downloadDir, dirEntry.Name()))
+				if err != nil {
+					return err
+				}
 			}
 		}
 	}
@@ -78,5 +85,6 @@ func init() {
 	rootCmd.AddCommand(downloadCmd)
 
 	downloadCmd.Flags().String("download-dir", "", "Download directory.")
+	downloadCmd.Flags().String("dpkg-status", "", "Path to /var/lib/dpkg/status file.")
 	_ = downloadCmd.MarkFlagRequired("download-dir")
 }
